Create Dockerfile
Let's try to make the thing useable by any CI system.

The whole idea is :
* You build the image which will compile a Godot version using all the V-Sekai patches
* Then you run the image by mapping a host-directory volume to /host
* Once ran, the host-directory will receive a 'godot' directory, which contain the executables and libraries that you can then tarball and release.

Now, this is only good for the first tests.
Afterwards, an image containing the whole set of V-Sekai repositories will need to be generated separately. Then separate Windows, Linux and Mac OS X builders will be used to generate different executables for the different platforms, using the same Docker image containing the code base as a basis.